refactor: Add refresh from `upstream/master` to "Making a Change" instructions.
Add instructions to explicitly refresh from `upstream/master` as the first step in the "Making a Change" instructions in the "How to Contribute" section. This will ensure that before any new changes are started a copy of the PMRoK Wiki content will be merged into the working branch.

Resolves: pmrok-chgmgt issue #21 suggestion #2
See also: https://github.com/Chingu-cohorts/pmrok-chgmgt/issues/21